App helper: Evaluate Gateway Perf
Idea of this app is to create an easy setup to evaluate
gateway performances. It is made of a single node that will
send as many packets as possible to the sink.

App must be flashed on a node, and a sink with dualmcu on a gateway.

Then the python script can be used to control the sending.
Backend ask the node to send N messages of size X as fast as it can.

Here is an example:
$ python gen_traffic_controler.py @<brocker setting file> --gateway_id <gateway_id>
(S)tart session, (Q)uit:Rx for different session
S
Starting a test session
Hom many packets?1000
Packet size?102
RX 1 packets in 0.000000s pps = -1.000000
...
RX 998 packets in 5.846717s pps = 170.694080
RX 999 packets in 5.863536s pps = 170.374999
RX 1000 packets in 5.864506s pps = 170.517353

It allows to watch CPU usage for example in parallel on the gateway.

Example above for example is on a rpi3 gw running docker images on edge tag
CPU for sink service: 8%
CPU for transport service: <20%